Disabled filesystem permission checks for windows. Windows filesystem permission is a bit different from UNIX like systems, and there is no good API in go to manipulate it. Nor is having these checks a requirement by the TUF spec.
This commit relies on build tags to inject the correct filesystem permssion
check, the Windows version always return "satisfied".
